Move mock package graph loader to TestSupport
(cherry picked from commit 4e4639b08ecbe69ffa418080b7a0169ac5254b7a)
